refactor: created common Input component which is used both inside adderesses and user details
