+++ src/components/client/UserAddress/components/common/Address.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
 
-import { AddressForm } from "@/components/client/UserAddress/components/Form/AddressForm";
+import { AddressForm } from "@/components/client/UserAddress/components/AddressForm";
 import { AddressTags } from "@/components/client/UserAddress/components/common/AddressTags";
-import { AddressInput } from "@/components/client/UserAddress/components/Form/AddressInput";
 import { AddressDelete } from "@/components/client/UserAddress/components/common/AddressDelete";
 import { AddressUpdate } from "@/components/client/UserAddress/components/common/AddressUpdate";
 import { AddressActions } from "@/components/client/UserAddress/components/common/AddressActions";
@@ -17,7 +16,6 @@ export default function Address({ children }: AddressProps) {
 
 Address.Tags = AddressTags;
 Address.Form = AddressForm;
-Address.Input = AddressInput;
 Address.Delete = AddressDelete;
 Address.Update = AddressUpdate;
 Address.Actions = AddressActions;
