+++ src/components/client/UserDetails/index.ts
@@ -1,7 +1,7 @@
 export { useUpdateUser } from "@/components/client/UserDetails/hooks/useUpdateUser";
 export { updateUserAction } from "@/components/client/UserDetails/actions/updateUserAction";
 
-export { default as Details } from "@/components/client/UserDetails/components/common/Details";
+export { default as Details } from "@/components/client/UserDetails/components/Details";
 export {
   hideEmail,
   hideName,
